04282015 UT Update/optimization of prod applciations
git-svn-id: https://www.etbmanager.org/svn/pharmadex/pharmadex/trunk@326 24bb7120-ba24-2248-a177-eda5b6c43647

+++ core/src/main/java/org/msh/pharmadex/dao/iface/PharmClassDAO.java
@@ -11,7 +11,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * To change this template use File | Settings | File Templates.
  */
 
-public interface PharmClassDAO extends JpaRepository<PharmClassif, Long> {
+public interface PharmClassDAO extends JpaRepository<PharmClassif, Integer> {
 
 
 }
